+++ src/platform_impl/orbital/window.rs
@@ -349,6 +349,16 @@ impl Window {
     }
 
     #[inline]
+    pub fn drag_resize_window(
+        &self,
+        _direction: window::ResizeDirection,
+    ) -> Result<(), error::ExternalError> {
+        Err(error::ExternalError::NotSupported(
+            error::NotSupportedError::new(),
+        ))
+    }
+
+    #[inline]
     pub fn set_cursor_hittest(&self, _hittest: bool) -> Result<(), error::ExternalError> {
         Err(error::ExternalError::NotSupported(
             error::NotSupportedError::new(),